refactor: simplifies folder structure
