+++ backend/services/player_records_service.py
@@ -1,4 +1,4 @@
-from services.records.records_service import RecordsService
+from backend.services.records_service import RecordsService
 
 
 class PlayerRecordsService:
